x86: dump the machine.yml file on boot
ARM and RISC-V machines are known at compile time and fully described
by their device tree files (DTS/DTB). On x86 machines however
resources are generally detected at runtime. This does not work well
for sel4cp which really needs to know where things are at compile
time.

This commit modifies the seL4 kernel to dump a machine.yml file on the
serial port with the physical addresses and sizes of various machine
resources. This machine.yml file can then be fed to the sel4cp tool to
build a static image for that machine.

This is just a proof of concept and a way to make progress on the port
of sel4cp to x86. A better implementation is planned as a standalone
tool.

Signed-off-by: Mathieu Mirmont <mat@neutrality.ch>